+++ backend/src/main/java/com/woowacourse/kkogkkog/auth/application/AuthService.java
@@ -3,7 +3,9 @@ package com.woowacourse.kkogkkog.auth.application;
 import com.woowacourse.kkogkkog.auth.application.dto.MemberUpdateResponse;
 import com.woowacourse.kkogkkog.auth.application.dto.TokenResponse;
 import com.woowacourse.kkogkkog.auth.support.JwtTokenProvider;
+import com.woowacourse.kkogkkog.infrastructure.application.GoogleClient;
 import com.woowacourse.kkogkkog.infrastructure.application.SlackClient;
+import com.woowacourse.kkogkkog.infrastructure.dto.GoogleUserInfo;
 import com.woowacourse.kkogkkog.infrastructure.dto.SlackUserInfo;
 import com.woowacourse.kkogkkog.infrastructure.dto.WorkspaceResponse;
 import com.woowacourse.kkogkkog.member.application.MemberService;
